+++ src/app/connection/types.rs
@@ -24,6 +24,7 @@ where
     sqlx::types::uuid::Uuid: sqlx::Type<D> + sqlx::Decode<'a, D>,
     sqlx::types::ipnetwork::IpNetwork: sqlx::Type<D> + sqlx::Decode<'a, D>,
     sqlx::types::mac_address::MacAddress: sqlx::Type<D> + sqlx::Decode<'a, D>,
+    sqlx::types::BitVec: sqlx::Type<D> + sqlx::Decode<'a, D>,
 
     sqlx_postgres::types::PgTimeTz: sqlx::Type<D> + sqlx::Decode<'a, D>,
 
@@ -47,6 +48,7 @@ where
     Vec<sqlx::types::uuid::Uuid>: sqlx::Type<D> + sqlx::Decode<'a, D>,
     Vec<sqlx::types::ipnetwork::IpNetwork>: sqlx::Type<D> + sqlx::Decode<'a, D>,
     Vec<sqlx::types::mac_address::MacAddress>: sqlx::Type<D> + sqlx::Decode<'a, D>,
+    Vec<sqlx::types::BitVec>: sqlx::Type<D> + sqlx::Decode<'a, D>,
 
     Vec<sqlx_postgres::types::PgTimeTz>: sqlx::Type<D> + sqlx::Decode<'a, D>,
 {
@@ -103,6 +105,9 @@ where
             return;
         }
         if Self::write_via_display::<sqlx::types::mac_address::MacAddress>(type_info, row, i, data) {
+            return;
+        }
+        if Self::write_via_debug::<sqlx::types::BitVec>(type_info, row, i, data) {
             return;
         }
         data.push("not supported".to_string());
@@ -202,6 +207,7 @@ where
     sqlx::types::uuid::Uuid: sqlx::Type<D> + sqlx::Decode<'a, D>,
     sqlx::types::ipnetwork::IpNetwork: sqlx::Type<D> + sqlx::Decode<'a, D>,
     sqlx::types::mac_address::MacAddress: sqlx::Type<D> + sqlx::Decode<'a, D>,
+    sqlx::types::BitVec: sqlx::Type<D> + sqlx::Decode<'a, D>,
     (): sqlx::Type<D> + sqlx::Decode<'a, D>,
     usize: ColumnIndex<R>,
 {
@@ -261,6 +267,9 @@ where
             return;
         }
         if Self::write_via_display::<sqlx::types::mac_address::MacAddress>(type_info, row, i, data) {
+            return;
+        }
+        if Self::write_via_debug::<sqlx::types::BitVec>(type_info, row, i, data) {
             return;
         }
         tracing::debug!("Type not supported: {:?}", type_info);